Created a reset modal for subreq (#352)
* Created a reset modal for subreq

* Fixed ts integration

* Fixed import order.

+++ src/components/Requirements/CompletedSubReqCourse.vue
@@ -1,5 +1,11 @@
 <template>
   <div class="completedsubreqcourse">
+    <reset-confirmation-modal
+      :isTestReq="isTransferCredit"
+      :reqName="courseTaken.code"
+      v-model="resetConfirmVisible"
+      @close-reset-modal="onResetConfirmClosed"
+    />
     <div class="completed-reqCourses-course-wrapper">
       <div class="separator"></div>
       <div class="completed-reqCourses-course-heading-wrapper">
@@ -33,6 +39,7 @@
 <script lang="ts">
 import Vue, { PropType } from 'vue';
 import ReqCourse from '@/components/Requirements/ReqCourse.vue';
+import ResetConfirmationModal from '@/components/Modals/ResetConfirmationModal.vue';
 import store from '@/store';
 import { deleteCourseFromSemesters } from '@/global-firestore-data';
 import getCurrentSeason, { getCurrentYear } from '@/utilities';
@@ -40,11 +47,14 @@ import getCurrentSeason, { getCurrentYear } from '@/utilities';
 const transferCreditColor = 'DA4A4A'; // Arbitrary color for transfer credit
 
 export default Vue.extend({
-  components: { ReqCourse },
+  components: { ReqCourse, ResetConfirmationModal },
   props: {
     subReqCourseId: { type: Number, required: true },
     courseTaken: { type: Object as PropType<CourseTaken>, required: true },
   },
+  data: () => ({
+    resetConfirmVisible: false,
+  }),
   computed: {
     semesters(): readonly FirestoreSemester[] {
       return store.state.semesters;
@@ -73,8 +83,11 @@ export default Vue.extend({
     },
   },
   methods: {
-    onReset() {
-      deleteCourseFromSemesters(this.courseTaken.uniqueId, this.$gtag);
+    onReset(): void {
+      this.resetConfirmVisible = true;
+    },
+    onResetConfirmClosed(isReset: boolean): void {
+      if (isReset) deleteCourseFromSemesters(this.courseTaken.uniqueId, this.$gtag);
     },
   },
 });
